chore: reorganize backend folder structure — Phase 4c (message functions)
Move message domain functions from app/lib/server/functions/ to
server/lib/messages/, updating all import paths and test mock keys.

extractUrlsFromMessageAST.spec.ts was not matched by any test-runner
glob before this move; it is now discovered by mocha via the new
server/lib/messages/ glob. Two of its assertions never matched the
implementation (protocol normalization) and are skipped with a TODO.
