+++ apps/meteor/.mocharc.js
@@ -15,6 +15,7 @@ module.exports = {
 	spec: [
 		'server/lib/callbacks.spec.ts',
 		'server/lib/cas/*.spec.ts',
+		'server/lib/messages/**/*.spec.ts',
 		'server/lib/ldap/*.spec.ts',
 		'server/lib/ldap/**/*.spec.ts',
 		'server/lib/dataExport/**/*.spec.ts',
