+++ apps/meteor/app/integrations/server/api/api.ts
@@ -17,9 +17,9 @@ import type { APIActionContext } from '../../../../server/api/router';
 import { loggerMiddleware } from '../../../../server/api/v1/middlewares/logger';
 import { metricsMiddleware } from '../../../../server/api/v1/middlewares/metrics';
 import { tracerSpanMiddleware } from '../../../../server/api/v1/middlewares/tracer';
+import type { WebhookResponseItem } from '../../../../server/lib/messages/processWebhookMessage';
+import { processWebhookMessage } from '../../../../server/lib/messages/processWebhookMessage';
 import { hasPermissionAsync } from '../../../authorization/server/functions/hasPermission';
-import type { WebhookResponseItem } from '../../../lib/server/functions/processWebhookMessage';
-import { processWebhookMessage } from '../../../lib/server/functions/processWebhookMessage';
 import { metrics } from '../../../metrics/server';
 import { settings } from '../../../settings/server';
 import { IsolatedVMScriptEngine } from '../lib/isolated-vm/isolated-vm';
