+++ apps/meteor/app/lib/server/methods/sendMessage.ts
@@ -11,12 +11,12 @@ import moment from 'moment';
 
 import { i18n } from '../../../../server/lib/i18n';
 import { SystemLogger } from '../../../../server/lib/logger/system';
+import { sendMessage } from '../../../../server/lib/messages/sendMessage';
 import { canSendMessageAsync } from '../../../authorization/server/functions/canSendMessage';
 import { hasPermissionAsync } from '../../../authorization/server/functions/hasPermission';
 import { applyAirGappedRestrictionsValidation } from '../../../license/server/airGappedRestrictionsWrapper';
 import { metrics } from '../../../metrics/server';
 import { settings } from '../../../settings/server';
-import { sendMessage } from '../functions/sendMessage';
 import { RateLimiter } from '../lib';
 /**
  *
