+++ apps/meteor/server/api/v1/chat.ts
@@ -33,8 +33,6 @@ import { Meteor } from 'meteor/meteor';
 import { roomAccessAttributes } from '../../../app/authorization/server';
 import { canAccessRoomAsync, canAccessRoomIdAsync } from '../../../app/authorization/server/functions/canAccessRoom';
 import { hasPermissionAsync } from '../../../app/authorization/server/functions/hasPermission';
-import { deleteMessageValidatingPermission } from '../../../app/lib/server/functions/deleteMessage';
-import { processWebhookMessage } from '../../../app/lib/server/functions/processWebhookMessage';
 import { getSingleMessage } from '../../../app/lib/server/methods/getSingleMessage';
 import { executeSendMessage } from '../../../app/lib/server/methods/sendMessage';
 import { executeUpdateMessage } from '../../../app/lib/server/methods/updateMessage';
@@ -46,6 +44,8 @@ import { settings } from '../../../app/settings/server';
 import { followMessage } from '../../../app/threads/server/methods/followMessage';
 import { unfollowMessage } from '../../../app/threads/server/methods/unfollowMessage';
 import { normalizeMessagesForUser } from '../../../app/utils/server/lib/normalizeMessagesForUser';
+import { deleteMessageValidatingPermission } from '../../lib/messages/deleteMessage';
+import { processWebhookMessage } from '../../lib/messages/processWebhookMessage';
 import { reportMessage } from '../../lib/moderation/reportMessage';
 import { ignoreUser } from '../../methods/ignoreUser';
 import { messageSearch } from '../../methods/messageSearch';
