+++ apps/meteor/server/api/v1/omnichannel/message.ts
@@ -13,11 +13,11 @@ import {
 
 import { API } from '../..';
 import { findGuest, findRoom, normalizeHttpHeaderData } from './lib/livechat';
-import { loadMessageHistory } from '../../../../app/lib/server/functions/loadMessageHistory';
 import { updateMessage, deleteMessage, sendMessage } from '../../../../app/livechat/server/lib/messages';
 import { settings } from '../../../../app/settings/server';
 import { normalizeMessageFileUpload } from '../../../../app/utils/server/functions/normalizeMessageFileUpload';
 import { callbacks } from '../../../lib/callbacks';
+import { loadMessageHistory } from '../../../lib/messages/loadMessageHistory';
 import { getPaginationItems } from '../../lib/getPaginationItems';
 import { isWidget } from '../../lib/isWidget';
