+++ apps/meteor/server/bridges/slack/RocketAdapter.ts
@@ -12,10 +12,10 @@ import { Meteor } from 'meteor/meteor';
 import _ from 'underscore';
 
 import { rocketLogger } from './logger';
-import { sendMessage } from '../../../app/lib/server/functions/sendMessage';
 import { settings } from '../../../app/settings/server';
 import { sleep } from '../../../lib/utils/sleep';
 import { callbacks } from '../../lib/callbacks';
+import { sendMessage } from '../../lib/messages/sendMessage';
 import { createRoom } from '../../lib/rooms/createRoom';
 import { setUserAvatar } from '../../lib/users/setUserAvatar';
