+++ apps/meteor/server/bridges/slack/SlackAdapter.ts
@@ -18,12 +18,12 @@ import { SlackAPI } from './SlackAPI';
 import { slackLogger } from './logger';
 import { saveRoomName, saveRoomTopic } from '../../../app/channel-settings/server';
 import { FileUpload } from '../../../app/file-upload/server';
-import { deleteMessage } from '../../../app/lib/server/functions/deleteMessage';
-import { sendMessage } from '../../../app/lib/server/functions/sendMessage';
-import { updateMessage } from '../../../app/lib/server/functions/updateMessage';
 import { executeSetReaction } from '../../../app/reactions/server/setReaction';
 import { settings } from '../../../app/settings/server';
 import { getUserAvatarURL } from '../../../app/utils/server/getUserAvatarURL';
+import { deleteMessage } from '../../lib/messages/deleteMessage';
+import { sendMessage } from '../../lib/messages/sendMessage';
+import { updateMessage } from '../../lib/messages/updateMessage';
 import { addUserToRoom } from '../../lib/rooms/addUserToRoom';
 import { archiveRoom } from '../../lib/rooms/archiveRoom';
 import { removeUserFromRoom } from '../../lib/rooms/removeUserFromRoom';
