+++ apps/meteor/server/lib/callbacks.ts
@@ -23,8 +23,8 @@ import type { Updater } from '@rocket.chat/models';
 import type { FilterOperators } from 'mongodb';
 
 import { Callbacks } from './callbacks/callbacksBase';
+import type { SendMessageOptions } from './messages/sendMessage';
 import type { ILoginAttempt } from '../../app/authentication/server/ILoginAttempt';
-import type { SendMessageOptions } from '../../app/lib/server/functions/sendMessage';
 import type { IBusinessHourBehavior } from '../../app/livechat/server/business-hour/AbstractBusinessHour';
 import type { CloseRoomParams } from '../../app/livechat/server/lib/localTypes';
