+++ apps/meteor/server/lib/messages/deleteMessage.ts
@@ -4,11 +4,15 @@ import { isThreadMessage, type AtLeast, type IMessage, type IRoom, type IThreadM
 import { Messages, Rooms, Uploads, Users, ReadReceipts, ReadReceiptsArchive, Subscriptions } from '@rocket.chat/models';
 import { Meteor } from 'meteor/meteor';
 
-import { callbacks } from '../../../../server/lib/callbacks';
-import { canDeleteMessageAsync } from '../../../authorization/server/functions/canDeleteMessage';
-import { FileUpload } from '../../../file-upload/server';
-import { settings } from '../../../settings/server';
-import { notifyOnRoomChangedById, notifyOnMessageChange, notifyOnSubscriptionChangedByRoomIdAndUserIds } from '../lib/notifyListener';
+import { canDeleteMessageAsync } from '../../../app/authorization/server/functions/canDeleteMessage';
+import { FileUpload } from '../../../app/file-upload/server';
+import {
+	notifyOnRoomChangedById,
+	notifyOnMessageChange,
+	notifyOnSubscriptionChangedByRoomIdAndUserIds,
+} from '../../../app/lib/server/lib/notifyListener';
+import { settings } from '../../../app/settings/server';
+import { callbacks } from '../callbacks';
 
 export const deleteMessageValidatingPermission = async (message: AtLeast<IMessage, '_id'>, userId: IUser['_id']): Promise<void> => {
 	if (!message?._id) {
