+++ apps/meteor/server/lib/messages/isTheLastMessage.ts
@@ -1,6 +1,6 @@
 import type { IMessage, IRoom, AtLeast } from '@rocket.chat/core-typings';
 
-import { settings } from '../../../settings/server';
+import { settings } from '../../../app/settings/server';
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export const isTheLastMessage = (room: AtLeast<IRoom, 'lastMessage'>, message: Pick<IMessage, '_id'>) =>
