+++ apps/meteor/server/lib/messages/processWebhookMessage.ts
@@ -4,12 +4,12 @@ import { Meteor } from 'meteor/meteor';
 import _ from 'underscore';
 
 import { sendMessage, validateMessage } from './sendMessage';
-import { ensureArray } from '../../../../lib/utils/arrayUtils';
-import { trim } from '../../../../lib/utils/stringUtils';
-import { SystemLogger } from '../../../../server/lib/logger/system';
-import { getRoomByNameOrIdWithOptionToJoin } from '../../../../server/lib/rooms/getRoomByNameOrIdWithOptionToJoin';
-import { validateRoomMessagePermissionsAsync } from '../../../authorization/server/functions/canSendMessage';
-import { settings } from '../../../settings/server';
+import { validateRoomMessagePermissionsAsync } from '../../../app/authorization/server/functions/canSendMessage';
+import { settings } from '../../../app/settings/server';
+import { ensureArray } from '../../../lib/utils/arrayUtils';
+import { trim } from '../../../lib/utils/stringUtils';
+import { SystemLogger } from '../logger/system';
+import { getRoomByNameOrIdWithOptionToJoin } from '../rooms/getRoomByNameOrIdWithOptionToJoin';
 
 type Payload = {
 	channel?: string | string[];
