+++ apps/meteor/server/methods/deleteFileMessage.ts
@@ -6,8 +6,8 @@ import { Meteor } from 'meteor/meteor';
 import type { DeleteResult } from 'mongodb';
 
 import { FileUpload } from '../../app/file-upload/server';
-import { deleteMessageValidatingPermission } from '../../app/lib/server/functions/deleteMessage';
 import { methodDeprecationLogger } from '../../app/lib/server/lib/deprecationWarningLogger';
+import { deleteMessageValidatingPermission } from '../lib/messages/deleteMessage';
 
 declare module '@rocket.chat/ddp-client' {
 	// eslint-disable-next-line @typescript-eslint/naming-convention
