+++ apps/meteor/server/methods/loadHistory.ts
@@ -6,8 +6,8 @@ import { Meteor } from 'meteor/meteor';
 
 import { canAccessRoomAsync, roomAccessAttributes } from '../../app/authorization/server';
 import { hasPermissionAsync } from '../../app/authorization/server/functions/hasPermission';
-import { loadMessageHistory } from '../../app/lib/server/functions/loadMessageHistory';
 import { settings } from '../../app/settings/server';
+import { loadMessageHistory } from '../lib/messages/loadMessageHistory';
 
 declare module '@rocket.chat/ddp-client' {
 	// eslint-disable-next-line @typescript-eslint/naming-convention
