+++ apps/meteor/server/services/media-call/service.ts
@@ -22,9 +22,9 @@ import { callStateToTranslationKey, getHistoryMessagePayload } from '@rocket.cha
 
 import { logger } from './logger';
 import { sendVoipPushNotification } from './push/sendVoipPushNotification';
-import { sendMessage } from '../../../app/lib/server/functions/sendMessage';
 import { settings } from '../../../app/settings/server';
 import { i18n } from '../../lib/i18n';
+import { sendMessage } from '../../lib/messages/sendMessage';
 import { createDirectMessage } from '../../methods/createDirectMessage';
 
 export class MediaCallService extends ServiceClassInternal implements IMediaCallService {
